requirements.txt
flask
groq
markdown
python-dotenv
